+++ apm-agent-plugins/apm-jms-plugin/src/test/java/co/elastic/apm/agent/jms/JmsInstrumentationIT.java
@@ -304,8 +304,8 @@ public class JmsInstrumentationIT extends AbstractInstrumentationTest {
         //noinspection ConstantConditions
         Id currentTraceId = tracer.currentTransaction().getTraceContext().getTraceId();
         assertThat(sendSpan.getTraceContext().getTraceId()).isEqualTo(currentTraceId);
-        assertThat(sendSpan.getContext().getMessage().getTopic().getName()).isNull();
-        assertThat(sendSpan.getContext().getMessage().getQueue().getName()).isEqualTo(queue.getQueueName());
+        assertThat(sendSpan.getContext().getMessage().getTopicName()).isNull();
+        assertThat(sendSpan.getContext().getMessage().getQueueName()).isEqualTo(queue.getQueueName());
 
         Id receiveTraceId = receiveSpan.getTraceContext().getTraceId();
         List<Transaction> receiveTransactions = reporter.getTransactions().stream().filter(transaction -> transaction.getTraceContext().getTraceId().equals(receiveTraceId)).collect(Collectors.toList());
@@ -316,8 +316,8 @@ public class JmsInstrumentationIT extends AbstractInstrumentationTest {
         if (sendToNoopSpan != null) {
             assertThat(sendToNoopSpan.getTraceContext().getTraceId()).isEqualTo(receiveTraceId);
             assertThat(sendToNoopSpan.getTraceContext().getParentId()).isEqualTo(receiveTransaction.getTraceContext().getId());
-            assertThat(sendToNoopSpan.getContext().getMessage().getTopic().getName()).isNull();
-            assertThat(sendToNoopSpan.getContext().getMessage().getQueue().getName()).isEqualTo("NOOP");
+            assertThat(sendToNoopSpan.getContext().getMessage().getTopicName()).isNull();
+            assertThat(sendToNoopSpan.getContext().getMessage().getQueueName()).isEqualTo("NOOP");
         }
     }
 
@@ -344,9 +344,9 @@ public class JmsInstrumentationIT extends AbstractInstrumentationTest {
         assertThat(spanName).endsWith(destinationName);
         boolean isQueue = spanName.contains("queue");
         if (isQueue) {
-            assertThat(sendSpan.getContext().getMessage().getQueue().getName()).isEqualTo(destinationName);
+            assertThat(sendSpan.getContext().getMessage().getQueueName()).isEqualTo(destinationName);
         } else {
-            assertThat(sendSpan.getContext().getMessage().getTopic().getName()).isEqualTo(destinationName);
+            assertThat(sendSpan.getContext().getMessage().getTopicName()).isEqualTo(destinationName);
         }
 
         //noinspection ConstantConditions
@@ -362,9 +362,9 @@ public class JmsInstrumentationIT extends AbstractInstrumentationTest {
             assertThat(receiveTransaction.getTraceContext().getParentId()).isEqualTo(sendSpan.getTraceContext().getId());
             assertThat(receiveTransaction.getType()).isEqualTo(MESSAGING_TYPE);
             if (isQueue) {
-                assertThat(receiveTransaction.getContext().getMessage().getQueue().getName()).isEqualTo(destinationName);
+                assertThat(receiveTransaction.getContext().getMessage().getQueueName()).isEqualTo(destinationName);
             } else {
-                assertThat(receiveTransaction.getContext().getMessage().getTopic().getName()).isEqualTo(destinationName);
+                assertThat(receiveTransaction.getContext().getMessage().getTopicName()).isEqualTo(destinationName);
             }
         }
     }
@@ -386,9 +386,9 @@ public class JmsInstrumentationIT extends AbstractInstrumentationTest {
         assertThat(spanName).endsWith(destinationName);
         boolean isQueue = spanName.contains("queue");
         if (isQueue) {
-            assertThat(sendInitialMessageSpan.getContext().getMessage().getQueue().getName()).isEqualTo(destinationName);
+            assertThat(sendInitialMessageSpan.getContext().getMessage().getQueueName()).isEqualTo(destinationName);
         } else {
-            assertThat(sendInitialMessageSpan.getContext().getMessage().getTopic().getName()).isEqualTo(destinationName);
+            assertThat(sendInitialMessageSpan.getContext().getMessage().getTopicName()).isEqualTo(destinationName);
         }
 
         //noinspection ConstantConditions
@@ -412,9 +412,9 @@ public class JmsInstrumentationIT extends AbstractInstrumentationTest {
             assertThat(receiveTransaction.getTraceContext().getTraceId()).isEqualTo(currentTraceId);
             assertThat(receiveTransaction.getTraceContext().getParentId()).isEqualTo(sendInitialMessageSpan.getTraceContext().getId());
             if (isQueue) {
-                assertThat(receiveTransaction.getContext().getMessage().getQueue().getName()).isEqualTo(destinationName);
+                assertThat(receiveTransaction.getContext().getMessage().getQueueName()).isEqualTo(destinationName);
             } else {
-                assertThat(receiveTransaction.getContext().getMessage().getTopic().getName()).isEqualTo(destinationName);
+                assertThat(receiveTransaction.getContext().getMessage().getTopicName()).isEqualTo(destinationName);
             }
             transactionId = receiveTransaction.getTraceContext().getId();
         }
@@ -425,7 +425,7 @@ public class JmsInstrumentationIT extends AbstractInstrumentationTest {
             assertThat(sendNoopSpan.getTraceContext().getTraceId()).isEqualTo(currentTraceId);
             // If both polling and handling transactions are captured, handling transaction would come second
             assertThat(sendNoopSpan.getTraceContext().getParentId()).isEqualTo(transactionId);
-            assertThat(sendNoopSpan.getContext().getMessage().getQueue().getName()).isEqualTo("NOOP");
+            assertThat(sendNoopSpan.getContext().getMessage().getQueueName()).isEqualTo("NOOP");
         }
     }