+++ CHANGELOG.md
@@ -14,6 +14,7 @@
  * Added `ElasticApmAttacher.attach(String propertiesLocation)` to specify a custom properties location
  * Logs message when `transaction_max_spans` has been exceeded (#849)
  * Report the number of affected rows by a SQL statement (UPDATE,DELETE,INSERT) in 'affected_rows' span attribute (#707)
+ * Report JMS destination as a span/transaction context field (#906)
 
 ## Bug Fixes
  * JMS creates polling transactions even when the API invocations return without a message